requirements.txt
discord-py-slash-command==3.0.3
discord.py==1.7.3
python-dotenv==0.19.2
mysql-connector-python==8.0.27
